Rename mockrdsclient.MockRDSClient to mocks.RDSClient
